SONAR-4093 Display number of tests covering a line

+++ sonar-plugin-api/src/main/java/org/sonar/api/test/Testable.java
@@ -30,6 +30,6 @@ public interface Testable extends Perspective {
 
   Collection<TestCase> testCasesCoveringLine(int line);
 
-  SortedSet<Integer> coveredLines();
+  SortedSet<Long> coveredLines();
 
 }
